separated Dropdown tests from DropdownInScroller tests
